requirements.txt
django==4.2.2
djangorestframework==3.14.0
django_extensions==3.1.5

google-auth==2.21.0
google-cloud==0.34.0
google-cloud-storage==2.10.0
aiohttp==3.8.4
asyncio==3.4.3
fireo==2.1.0
firebase-admin==6.2.0
httpx==0.23.3